// d/src/interfaces/libs/http.ts
export enum HttpMethod {
    GET = 'GET',
    POST = 'POST',
    OPTIONS = 'OPTIONS',
    PUT = 'PUT',
    DELETE = 'DELETE',
    PATCH = 'PATCH'
}

export enum HttpProtocol {
    Http = 'http',
    Https = 'https'
}

export interface HttpServiceResponseResult {
    data?: any;
    message?: any;
    headers?: IncomingMessage['headers'];
    statusCode?: HttpStatusCode;
    code?: HttpStatusCode;
    statusMessage?: string;
}

export type HttpServiceResponse = [Error & HttpServiceResponseResult, HttpServiceResponseResult];
// ...
export interface MakeApiCallOps {
    method?: HttpMethod;
    data?: unknown;
}